+++ entrypoint.sh
@@ -68,10 +68,19 @@ REMOTE_PORT_FILE=$IMAGE_ROOT_PARENT/remote_port
 REMOTE_PORT=""
 
 PARAMETER_FILE_BUILD_FAILED=$IMAGE_ROOT_PARENT/build_failed
+PARAMETER_FILE_POD_ERROR=$IMAGE_ROOT_PARENT/pod_error
+
+# NOTE: these are locations within the remote container
+SIGNAL_FILE_REMOTE_EXITING=/mnt/image/remote_exiting
+SIGNAL_FILE_REMOTE_ERROR=image/remote_error
+
+# Number of consecutive heartbeat failures before assuming remote job is no longer running
+# Default is 12 (1 minute if checking every 5 seconds) - set in IMS configuration map
+HEARTBEAT_ERROR_LIMIT=${REMOTE_HEARTBEAT_TIMEOUT_COUNT:-12}
 
 function wait_for_ready {
     echo "Waiting for $SIGNAL_FILE_READY flag"
-    until [ -f "$SIGNAL_FILE_READY" ]
+    until [[ -f ${SIGNAL_FILE_READY} ]]
     do
         echo "Waiting for $SIGNAL_FILE_READY to exist before starting ssh environment $(date)"
         sleep 5;
@@ -79,7 +88,7 @@ function wait_for_ready {
 }
 
 function wait_for_local_complete {
-    until [ -f "$SIGNAL_FILE_COMPLETE" ] || [ -f "$SIGNAL_FILE_FAILED" ]
+    until [[ -f ${SIGNAL_FILE_COMPLETE} || -f ${SIGNAL_FILE_FAILED} ]]
     do
         sleep 5;
     done
@@ -87,14 +96,40 @@ function wait_for_local_complete {
 
 function wait_for_remote_complete {
     # Loop forever until the user is done
+    HEARTBEAT_ERROR_COUNT=0
     while [ true ]
     do
         # Look for the exiting flag in the remote job
-        ssh -o StrictHostKeyChecking=no root@${REMOTE_BUILD_NODE} "podman cp ims-${IMS_JOB_ID}:/mnt/image/remote_exiting /tmp/ims_${IMS_JOB_ID}"
-        rc=$?
-        if [ "$rc" -eq "0" ]; then
-            # a return value of 0 indicates file is present - remote complete
+        if ssh -o StrictHostKeyChecking=no "root@${REMOTE_BUILD_NODE}" "test -f /tmp/ims_${IMS_JOB_ID}/${SIGNAL_FILE_REMOTE_EXITING}"; then
             return 0
+        fi
+
+        # Look for the error flag in the remote job
+        if ssh -o StrictHostKeyChecking=no "root@${REMOTE_BUILD_NODE}" "test -f /tmp/ims_${IMS_JOB_ID}/${SIGNAL_FILE_REMOTE_ERROR}"; then
+            # a return value of 0 indicates file is present - remote complete
+            echo "Remote job reported a container error - exiting wait loop"
+            touch "${SIGNAL_FILE_FAILED}" "${PARAMETER_FILE_POD_ERROR}"
+            return 1
+        fi
+
+        # make sure the remote job is still running
+        ssh -o StrictHostKeyChecking=no "root@${REMOTE_BUILD_NODE}" "podman exec ims-${IMS_JOB_ID} echo heartbeat">/dev/null 2>&1
+        # a return value of 0 indicates the container is running
+        if [[ $? -ne 0 ]]; then
+            # Heartbeat failed - not sure if this is a pod failure or temporary network issue
+            if [[ ${HEARTBEAT_ERROR_COUNT} -ge ${HEARTBEAT_ERROR_LIMIT} ]]; then
+                # After 12 consecutive errors (1 minute) assume the remote job is no longer running
+                echo "Remote job is no longer accessible after ${HEARTBEAT_ERROR_LIMIT} consecutive errors - exiting wait loop"
+                touch "${SIGNAL_FILE_FAILED}" "${PARAMETER_FILE_POD_ERROR}"
+                return 1
+            fi
+
+            # Increment error count and try again
+            HEARTBEAT_ERROR_COUNT=$((HEARTBEAT_ERROR_COUNT+1))
+            echo "Remote container heartbeat failed - error count: ${HEARTBEAT_ERROR_COUNT}"
+        else
+            # Heartbeat succeeded - reset error count
+            HEARTBEAT_ERROR_COUNT=0
         fi
 
         sleep 5;
@@ -103,40 +138,45 @@ function wait_for_remote_complete {
 
 function wait_for_complete {
 
-    echo "To mark this shell as successful, touch the file \"$SIGNAL_FILE_COMPLETE\"."
-    echo "To mark this shell as failed, touch the file \"$SIGNAL_FILE_FAILED\"."
+    echo "To mark this shell as successful, touch the file \"${SIGNAL_FILE_COMPLETE}\"."
+    echo "To mark this shell as failed, touch the file \"${SIGNAL_FILE_FAILED}\"."
     echo "Waiting for User to mark this shell as either successful or failed."
 
     # If this is a remote build, we need to check the remote job for completion
-    if [[ -n "${REMOTE_BUILD_NODE}" ]]; then
+    RC=0
+    if [[ -n ${REMOTE_BUILD_NODE} ]]; then
+        # pass along the return code from the wait_for_remote_complete function in case
+        # we need to signal an error with the remote pod
         wait_for_remote_complete
+        RC=$?
     else
         wait_for_local_complete
     fi
 
     # check the results of the run now that the user shell has exited
-    if [ -f "$SIGNAL_FILE_FAILED" ]
+    if [[ -f ${SIGNAL_FILE_FAILED} ]]
     then
-        echo "$SIGNAL_FILE_FAILED exists; Shell was marked failed."
-    elif [ -f "$SIGNAL_FILE_COMPLETE" ]
+        echo "${SIGNAL_FILE_FAILED} exists; Shell was marked failed."
+    elif [[ -f ${SIGNAL_FILE_COMPLETE} ]]
     then
-        echo "$SIGNAL_FILE_COMPLETE exists; Shell was marked successful."
+        echo "${SIGNAL_FILE_COMPLETE} exists; Shell was marked successful."
     fi
 
-    if [ -f "$SIGNAL_FILE_COMPLETE" ]
+    if [[ -f ${SIGNAL_FILE_COMPLETE} ]]
     then
         # Remove the complete file now that we're done
-        echo "Removing $SIGNAL_FILE_COMPLETE"
-        rm "$SIGNAL_FILE_COMPLETE"
+        echo "Removing ${SIGNAL_FILE_COMPLETE}"
+        rm "${SIGNAL_FILE_COMPLETE}"
     fi
 
     echo "Exiting ssh environment"
+    return $RC
 }
 
 function signal_exiting {
     # Let the buildenv-sidecar container know that we're exiting
-    echo "Touching $SIGNAL_FILE_EXITING flag"
-    touch "$SIGNAL_FILE_EXITING"
+    echo "Touching ${SIGNAL_FILE_EXITING} flag"
+    touch "${SIGNAL_FILE_EXITING}"
 }
 
 function run_user_shell {
@@ -151,20 +191,20 @@ function run_user_shell {
     echo "SetEnv IMS_JOB_ID=$IMS_JOB_ID IMS_ARCH=$BUILD_ARCH IMS_DKMS_ENABLED=$JOB_ENABLE_DKMS REMOTE_BUILD_NODE=$REMOTE_BUILD_NODE" >> "$SSHD_CONFIG_FILE"
 
     # Set up forwarding to remote node if needed
-    if [[ -n "${REMOTE_BUILD_NODE}" ]]; then
+    if [[ -n ${REMOTE_BUILD_NODE} ]]; then
         # if the remote port file does not exist, bail
-        if [ ! -f ${REMOTE_PORT_FILE} ]; then
+        if [[ ! -f ${REMOTE_PORT_FILE} ]]; then
             echo "ERROR: file with remote port missing - can not proceed with remote job!"
-            touch "$SIGNAL_FILE_FAILED"
+            touch "${SIGNAL_FILE_FAILED}"
             signal_exiting
-            return
+            return 1
         fi
 
         # get remote port file into env var REMOTE_PORT
         REMOTE_PORT=$(cat ${REMOTE_PORT_FILE})
 
         # set up cleanup in case this exits unexpectedly
-        trap clean_exit SIGTERM SIGINT
+        trap clean_exit SIGTERM SIGINT SIGKILL
 
         # prepare the ssh keys to access the remote node
         mkdir -p ~/.ssh
@@ -179,29 +219,29 @@ function run_user_shell {
     fi
 
     # Setup SSH jail
-    if [ "$SSH_JAIL" = "True" ]
+    if [[ $SSH_JAIL == "True" ]];
     then
         chmod 755 "$IMAGE_ROOT_PARENT"
         chown root:root "$IMAGE_ROOT_PARENT"
         chown root:root "$IMAGE_ROOT_PARENT/image-root"
 
         # if this is a remote job, chrootDir is set up on remote config
-        if [[ -z "${REMOTE_BUILD_NODE}" ]]; then
-            echo "Match User root" >> "$SSHD_CONFIG_FILE"
-            echo "ChrootDirectory $IMAGE_ROOT_PARENT/image-root" >> "$SSHD_CONFIG_FILE"
+        if [[ -z ${REMOTE_BUILD_NODE} ]]; then
+            echo "Match User root" >> "${SSHD_CONFIG_FILE}"
+            echo "ChrootDirectory ${IMAGE_ROOT_PARENT}/image-root" >> "${SSHD_CONFIG_FILE}"
 
             # If this is not a remote job, change location of complete files
-            SIGNAL_FILE_COMPLETE=$IMAGE_ROOT_PARENT/image-root/tmp/complete
-            SIGNAL_FILE_FAILED=$IMAGE_ROOT_PARENT/image-root/tmp/failed
+            SIGNAL_FILE_COMPLETE=${IMAGE_ROOT_PARENT}/image-root/tmp/complete
+            SIGNAL_FILE_FAILED=${IMAGE_ROOT_PARENT}/image-root/tmp/failed
         fi
     fi
 
     # If setting up for dkms permissions, do that now
-    echo "JOB_ENABLE_DKMS: $JOB_ENABLE_DKMS"
-    local is_dkms=$(echo $JOB_ENABLE_DKMS | tr '[:upper:]' '[:lower:]')
-    echo "is_dkms=$is_dkms"
-    if [ "$is_dkms" = "true" ]; then
-        if [[ -n "${REMOTE_BUILD_NODE}" ]]; then
+    echo "JOB_ENABLE_DKMS: ${JOB_ENABLE_DKMS}"
+    local is_dkms=$(echo "${JOB_ENABLE_DKMS}" | tr '[:upper:]' '[:lower:]')
+    echo "is_dkms=${is_dkms}"
+    if [[ ${is_dkms} == "true" ]]; then
+        if [[ -n ${REMOTE_BUILD_NODE} ]]; then
             echo " dkms mounts not set in sshd pod for remote jobs"
         else
             if mount -t sysfs /sysfs /mnt/image/image-root/sys; then
@@ -225,11 +265,11 @@ function run_user_shell {
     fi
 
     # If setting up for arm64 emulation, do that now
-    echo "Checking build architecture: $BUILD_ARCH"
-    if [ "$BUILD_ARCH" == "aarch64" ]; then
+    echo "Checking build architecture: ${BUILD_ARCH}"
+    if [[ ${BUILD_ARCH} == "aarch64" ]]; then
         echo "Build architecture is aarch64"
-        # Register qemu-aarch64-static to act as an arm interpreter for arm builds 
-        if [ ! -d /proc/sys/fs/binfmt_misc ] ; then
+        # Register qemu-aarch64-static to act as an arm interpreter for arm builds
+        if [[ ! -d /proc/sys/fs/binfmt_misc ]]; then
             echo "- binfmt_misc does not appear to be loaded or isn't built in."
             echo "  Trying to load it..."
             if ! modprobe binfmt_misc ; then
@@ -239,7 +279,7 @@ function run_user_shell {
         fi
 
         # mount the emulation filesystem
-        if [ ! -f /proc/sys/fs/binfmt_misc/register ] ; then
+        if [[ ! -f /proc/sys/fs/binfmt_misc/register ]] ; then
             echo "- The binfmt_misc filesystem does not appear to be mounted."
             echo "  Trying to mount it..."
             if ! mount binfmt_misc -t binfmt_misc /proc/sys/fs/binfmt_misc ; then
@@ -249,7 +289,7 @@ function run_user_shell {
         fi
 
         # register qemu for aarch64 images 
-        if [ ! -f /proc/sys/fs/binfmt_misc/qemu-aarch64 ] ; then
+        if [[ ! -f /proc/sys/fs/binfmt_misc/qemu-aarch64 ]] ; then
             echo "- Setting up QEMU for ARM64"
             echo ":qemu-aarch64:M::\x7f\x45\x4c\x46\x02\x01\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x02\x00\xb7\x00:\xff\xff\xff\xff\xff\xff\xff\x00\xff\xff\xff\xff\xff\xff\xff\xff\xfe\xff\xff\xff:/usr/bin/qemu-aarch64-static:F" >> /proc/sys/fs/binfmt_misc/register
         fi
@@ -259,29 +299,33 @@ function run_user_shell {
     chown -R root:root /etc/cray/ims
 
     # Start the SSH server daemon or set up port forwarding
-    if [[ -z "${REMOTE_BUILD_NODE}" ]]; then
+    if [[ -z ${REMOTE_BUILD_NODE} ]]; then
         ssh-keygen -A
         /usr/sbin/sshd $SSHD_OPTIONS
     fi
 
     # Perform any other bootstrapping tasks here or run a script
     # located in the build environment
-    if [ ! -z $CUSTOMIZATION_SCRIPT ]
-    then
-        . $CUSTOMIZATION_SCRIPT
+    if [[ ! -z ${CUSTOMIZATION_SCRIPT} ]]; then
+        . ${CUSTOMIZATION_SCRIPT}
     fi
 
-    # Enter wait loop for $SIGNAL_FILE_COMPLETE to show up
+    # Enter wait loop for $SIGNAL_FILE_COMPLETE to show up - pass along return code
+    # This will mark the k8s job in an error state if the remote job exits with an error
     wait_for_complete
+    RC=$?
 
     # If this is a remote customize build, we need to pull the results back
     # from the remote node.
-    if [[ -n "${REMOTE_BUILD_NODE}" ]]; then
+    if [[ -n ${REMOTE_BUILD_NODE} ]]; then
         fetch_remote_artifacts
     fi
 
     # Let the buildenv-sidecar container know that we're exiting
     signal_exiting
+
+    # pass along the return code from the wait_for_complete function
+    return $RC
 }
 
 function clean_exit {
@@ -289,7 +333,7 @@ function clean_exit {
     echo "Abrupt exiting..."
 
     # remote container should still be running - need to kill it
-    ssh -o StrictHostKeyChecking=no root@${REMOTE_BUILD_NODE} "podman stop ims-${IMS_JOB_ID}"
+    ssh -o StrictHostKeyChecking=no "root@${REMOTE_BUILD_NODE}" "podman stop ims-${IMS_JOB_ID}"
 
     # now that container has stopped, we can proceed with the normal cleanup
     clean_remote_node
@@ -300,17 +344,17 @@ function clean_exit {
 function clean_remote_node {
     # delete artifacts off of remote host
     # NOTE: need to prune the anonymous volume explicitly to free up the space
-    ssh -o StrictHostKeyChecking=no root@${REMOTE_BUILD_NODE} "rm -rf /tmp/ims_${IMS_JOB_ID}/"
-    ssh -o StrictHostKeyChecking=no root@${REMOTE_BUILD_NODE} "podman rm ims-${IMS_JOB_ID}"
-    ssh -o StrictHostKeyChecking=no root@${REMOTE_BUILD_NODE} "podman rmi ims-remote-${IMS_JOB_ID}:1.0.0"
-    ssh -o StrictHostKeyChecking=no root@${REMOTE_BUILD_NODE} "podman volume prune -f"
+    ssh -o StrictHostKeyChecking=no "root@${REMOTE_BUILD_NODE}" "rm -rf /tmp/ims_${IMS_JOB_ID}/"
+    ssh -o StrictHostKeyChecking=no "root@${REMOTE_BUILD_NODE}" "podman rm ims-${IMS_JOB_ID}"
+    ssh -o StrictHostKeyChecking=no "root@${REMOTE_BUILD_NODE}" "podman rmi ims-remote-${IMS_JOB_ID}:1.0.0"
+    ssh -o StrictHostKeyChecking=no "root@${REMOTE_BUILD_NODE}" "podman volume prune -f"
 }
 
 function fetch_remote_artifacts {
     # check the results of the build
-    ssh -o StrictHostKeyChecking=no root@${REMOTE_BUILD_NODE} "podman cp ims-${IMS_JOB_ID}:/mnt/image/complete /tmp/ims_${IMS_JOB_ID}/"
+    ssh -o StrictHostKeyChecking=no "root@${REMOTE_BUILD_NODE}" "podman cp ims-${IMS_JOB_ID}:/mnt/image/complete /tmp/ims_${IMS_JOB_ID}/"
     rc=$?
-    if [ "$rc" -ne "0" ]; then
+    if [[ $rc -ne 0 ]]; then
         # Failed rc indicates file not present
         echo "ERROR: Error reported from customize job."
 
@@ -318,27 +362,27 @@ function fetch_remote_artifacts {
         clean_remote_node
 
         # signal we are done
-        echo "Touching failed flag: $SIGNAL_FILE_FAILED"
-        touch $SIGNAL_FILE_FAILED
+        echo "Touching failed flag: ${SIGNAL_FILE_FAILED}"
+        touch "${SIGNAL_FILE_FAILED}"
     else
         echo "Remote job succeeded - fetching remote artifacts..."
 
         # copy image files from remote machine to job pod
-        scp -o StrictHostKeyChecking=no root@${REMOTE_BUILD_NODE}:/tmp/ims_${IMS_JOB_ID}/image/* ${IMAGE_ROOT_PARENT}
+        scp -o StrictHostKeyChecking=no "root@${REMOTE_BUILD_NODE}:/tmp/ims_${IMS_JOB_ID}/image/*" "${IMAGE_ROOT_PARENT}"
 
         # make sure the remote node artifacts are cleaned up
         clean_remote_node
 
         # signal we are done
-        touch $SIGNAL_FILE_COMPLETE
+        touch "${SIGNAL_FILE_COMPLETE}"
     fi
 }
 
 function should_run_user_shell {
     case "$IMS_ACTION" in
         create)
-            if [ -f "$PARAMETER_FILE_BUILD_FAILED" ]; then
-                if [[ $(echo "$ENABLE_DEBUG" | tr [:upper:] [:lower:]) = "true" ]]; then
+            if [[ -f $PARAMETER_FILE_BUILD_FAILED ]]; then
+                if [[ $(echo "${ENABLE_DEBUG}" | tr [:upper:] [:lower:]) = "true" ]]; then
                     echo "Running user shell for failed create action"
                     return 0
                 else
@@ -355,7 +399,7 @@ function should_run_user_shell {
             return 0
             ;;
          *)
-            echo "Unknown IMS Action: $IMS_ACTION. Not running user shell."
+            echo "Unknown IMS Action: ${IMS_ACTION}. Not running user shell."
             return 1
             ;;
     esac
@@ -364,5 +408,9 @@ function should_run_user_shell {
 
 # Do we need to present the user with a SSH shell?
 if should_run_user_shell; then
+  # A non-zero return code indicates the user shell failed - used to signal that
+  # the remote job container failed in an unexpected way. This will mark the k8s job
+  # in an error state.
   run_user_shell
+  exit $?
 fi