+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- CASMCMS-9560 - better checking for remote build node port availability.
+- CASMTRIAGE-8568 - better handling of remote build node container failures.
 
 ## [1.14.0] - 2025-10-01
 ## Changed